Created AsyncStorage wrapper so it is mockable without replacing all of react-native, mocking it in test setup file.
