+++ app/lib/AsyncStorage.js
@@ -1,0 +1,4 @@
+import { AsyncStorage } from 'react-native'
+
+/** Wrapper class so AsyncStorage can be mockable in tests without mocking all of react-native.  */
+export default AsyncStorage
